+++ runtime/drivers/druid/olap.go
@@ -17,7 +17,7 @@ func (c *connection) WithConnection(ctx context.Context, priority int, fn driver
 	panic("not implemented")
 }
 
-func (c *connection) WithRaw(ctx context.Context, priority int, fn drivers.RawFunc) error {
+func (c *connection) WithRaw(ctx context.Context, priority int, fn drivers.WithRawFunc) error {
 	panic("not implemented")
 }
