+++ runtime/drivers/sql_store.go
@@ -11,6 +11,9 @@ type Field struct {
 
 type Schema []Field
 
+// SQLStore is implemented by drivers capable of running sql queries and generating an iterator to consume results.
+// In future the results can be produced in other formats like arrow as well.
+// This is kept differnt from OLAP store since results can be produced in other formats in future and drivers like gcs, athena may not produce sql.Rows.
 type SQLStore interface {
 	Exec(ctx context.Context, src *DatabaseSource) (RowIterator, error)
 }
@@ -24,4 +27,7 @@ type RowIterator interface {
 	Next(ctx context.Context) ([]any, error)
 	// Close closes the iterator and frees resources
 	Close() error
+	// Size returns size of data downloaded in unit.
+	// Returns 0,false if not able to compute size in given unit
+	Size(unit ProgressUnit) (uint64, bool)
 }
