+++ web-common/src/features/sources/sourceUtils.ts
@@ -20,7 +20,11 @@ export function compileCreateSourceYAML(
 ) {
   const topLineComment = `# Visit https://docs.rilldata.com/ to learn more about Rill code files.`;
 
-  if (connectorName !== "local_file" && connectorName !== "motherduck" && connectorName !== "bigquery") {
+  if (
+    connectorName !== "local_file" &&
+    connectorName !== "motherduck" &&
+    connectorName !== "bigquery"
+  ) {
     values.uri = values.path;
     delete values.path;
   }
